requirements.txt
albumentations>=1.4.0
neptune>=1.7.0
numpy>=1.23
omegaconf>=2.3
pycocotools>=2.0.7
python-dotenv>=1.0
PyYAML>=6.0
torch>=2.2
torchvision>=0.17
transformers>=4.38
tqdm>=4.66
